requirements.txt
# PyTorch (MPS 지원 - MacBook M4)
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Deep Learning Models
timm>=0.9.0
ultralytics>=8.0.0

# Image Processing & Augmentation
albumentations>=1.3.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Visualization & Interpretability
grad-cam>=1.3.0

# API Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
anthropic>=0.18.0
python-dotenv>=1.0.0
reportlab>=4.0.0

# AWS Integration
boto3>=1.28.0

# Experiment Tracking
wandb>=0.15.0

# Model Export & Optimization
onnx>=1.14.0
onnxruntime>=1.16.0

# Utilities
tqdm>=4.66.0
numpy>=1.24.0
matplotlib>=3.7.0
